Add role in group

+++ lib/presentation/features/market/groups/group_detail/bloc/group_detail_bloc.dart
@@ -4,6 +4,8 @@ import 'package:plan_meal_app/data/model/group.dart';
 import 'package:plan_meal_app/data/model/group_member.dart';
 import 'package:plan_meal_app/data/repositories/abstract/group_repository.dart';
 import 'package:plan_meal_app/domain/entities/member_entity.dart';
+import 'package:plan_meal_app/domain/preference_utils.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
 part 'group_detail_event.dart';
 
@@ -25,6 +27,7 @@ class GroupDetailBloc extends Bloc<GroupDetailEvent, GroupDetailState> {
       emit(GroupDetailLoading());
       List<GroupMember> groupMemberList =
           await groupRepository.getMemberListByGroupId(groupId: event.groupId);
+      int userId = int.parse(PreferenceUtils.getString("userId") ?? "-1");
       List<MemberEntity> memberEntityList = [];
       for (var member in groupMemberList) {
         var user = member.user;
@@ -41,7 +44,8 @@ class GroupDetailBloc extends Bloc<GroupDetailEvent, GroupDetailState> {
             userId: member.userId ?? 0);
         memberEntityList.add(memberEntity);
       }
-      emit(GroupDetailHasMember(listMember: memberEntityList));
+      var memberResult = memberEntityList.firstWhere((element) => element.userId == userId);
+      emit(GroupDetailHasMember(listMember: memberEntityList, isAdmin: memberResult.isAdmin));
     } catch (exception) {
       emit(GroupDetailError(error: exception.toString()));
     }
@@ -51,21 +55,26 @@ class GroupDetailBloc extends Bloc<GroupDetailEvent, GroupDetailState> {
       GroupDetailRemoveMemberEvent event,
       Emitter<GroupDetailState> emit) async {
     emit(GroupDetailWaiting());
-    var result = await groupRepository.removeMember(
-        event.memberId.toString(), event.groupId.toString());
-    if (result == "201") {
-      List<MemberEntity> memberList = [];
-      memberList.addAll(event.memberList);
-      memberList.removeWhere((element) => element.userId == event.memberId);
-      emit(GroupDetailFinished());
-      emit(GroupDetailHasMember(listMember: memberList));
+    if (event.isAdmin) {
+      var result = await groupRepository.removeMember(
+          event.memberId.toString(), event.groupId.toString());
+      if (result == "201") {
+        List<MemberEntity> memberList = [];
+        memberList.addAll(event.memberList);
+        memberList.removeWhere((element) => element.userId == event.memberId);
+        emit(GroupDetailFinished());
+        emit(GroupDetailHasMember(listMember: memberList, isAdmin: event.isAdmin));
+      }
+      emit(GroupDetailHasMember(listMember: event.memberList, isAdmin: event.isAdmin));
     }
-    emit(GroupDetailHasMember(listMember: event.memberList));
   }
 
   Future<void> _onGroupDetailDeleteGroupEvent(GroupDetailDeleteGroupEvent event, Emitter<GroupDetailState> emit) async {
     String result = await groupRepository.deleteGroup(event.groupId.toString());
     if (result == "200") {
+      var prefs = await SharedPreferences.getInstance();
+      prefs.remove("groupId");
+      prefs.remove("groupName");
       emit(GroupDetailDeleted());
     }
   }
